+++ ifs-resources/src/test/java/org/innovateuk/ifs/competition/builder/CompetitionOpenKeyStatisticsResourceBuilderTest.java
@@ -13,12 +13,12 @@ public class CompetitionOpenKeyStatisticsResourceBuilderTest {
     @Test
     public void buildOne() {
 
-        long expectedAssessorsInvited = 1L;
-        long expectedAssessorsAccepted = 2L;
-        long expectedApplicationsPerAssessor = 3L;
-        long expectedApplicationsStarted = 4L;
-        long expectedApplicationsPastHalf = 5L;
-        long expectedApplicationsSubmitted = 6L;
+        int expectedAssessorsInvited = 1;
+        int expectedAssessorsAccepted = 2;
+        int expectedApplicationsPerAssessor = 3;
+        int expectedApplicationsStarted = 4;
+        int expectedApplicationsPastHalf = 5;
+        int expectedApplicationsSubmitted = 6;
 
         CompetitionOpenKeyStatisticsResource keyStatisticsResource = newCompetitionOpenKeyStatisticsResource()
                 .withAssessorsInvited(expectedAssessorsInvited)
@@ -40,12 +40,12 @@ public class CompetitionOpenKeyStatisticsResourceBuilderTest {
 
     @Test
     public void buildMany() {
-        Long[] expectedAssessorsInviteds = {1L, 11L};
-        Long[] expectedAssessorsAccepteds = {2L, 12L};
-        Long[] expectedApplicationsPerAssessors = {3L, 13L};
-        Long[] expectedApplicationsStarteds = {4L, 14L};
-        Long[] expectedApplicationsPastHalfs = {5L, 15L};
-        Long[] expectedApplicationsSubmitteds = {6L, 16L};
+        Integer[] expectedAssessorsInviteds = {1, 11};
+        Integer[] expectedAssessorsAccepteds = {2, 12};
+        Integer[] expectedApplicationsPerAssessors = {3, 13};
+        Integer[] expectedApplicationsStarteds = {4, 14};
+        Integer[] expectedApplicationsPastHalfs = {5, 15};
+        Integer[] expectedApplicationsSubmitteds = {6, 16};
 
         List<CompetitionOpenKeyStatisticsResource> keyStatisticsResources = newCompetitionOpenKeyStatisticsResource()
                 .withAssessorsInvited(expectedAssessorsInviteds)
@@ -57,12 +57,12 @@ public class CompetitionOpenKeyStatisticsResourceBuilderTest {
                 .build(2);
 
         for (int i = 0; i < keyStatisticsResources.size(); i++) {
-            assertEquals((long) expectedAssessorsInviteds[i], keyStatisticsResources.get(i).getAssessorsInvited());
-            assertEquals((long) expectedAssessorsAccepteds[i], keyStatisticsResources.get(i).getAssessorsAccepted());
-            assertEquals((long) expectedApplicationsPerAssessors[i], keyStatisticsResources.get(i).getApplicationsPerAssessor());
-            assertEquals((long) expectedApplicationsStarteds[i], keyStatisticsResources.get(i).getApplicationsStarted());
-            assertEquals((long) expectedApplicationsPastHalfs[i], keyStatisticsResources.get(i).getApplicationsPastHalf());
-            assertEquals((long) expectedApplicationsSubmitteds[i], keyStatisticsResources.get(i).getApplicationsSubmitted());
+            assertEquals((int) expectedAssessorsInviteds[i], keyStatisticsResources.get(i).getAssessorsInvited());
+            assertEquals((int) expectedAssessorsAccepteds[i], keyStatisticsResources.get(i).getAssessorsAccepted());
+            assertEquals((int) expectedApplicationsPerAssessors[i], keyStatisticsResources.get(i).getApplicationsPerAssessor());
+            assertEquals((int) expectedApplicationsStarteds[i], keyStatisticsResources.get(i).getApplicationsStarted());
+            assertEquals((int) expectedApplicationsPastHalfs[i], keyStatisticsResources.get(i).getApplicationsPastHalf());
+            assertEquals((int) expectedApplicationsSubmitteds[i], keyStatisticsResources.get(i).getApplicationsSubmitted());
         }
 
     }
